+++ packages/miniprogram-render/src/tree/tree.js
@@ -1,4 +1,5 @@
 const QuerySelector = require('./query-selector')
+const tool = require('../util/tool')
 
 /**
  * 遍历 dom 树，收集类和标签对应的节点列表
@@ -98,6 +99,12 @@ class Tree {
     getByTagName(tagName, node) {
         const cache = {}
         walkDomTree(node || this.root, cache)
+
+        if (tool.checkIsWxComponent(tagName.toLowerCase(), false)) {
+            // 内置组件
+            tagName = tagName.toLowerCase().slice(3)
+            return (cache.tagMap['WX-COMPONENT'] || []).filter(findNode => findNode.behavior === tagName)
+        }
 
         return cache.tagMap[tagName.toUpperCase()] || []
     }