+++ packages/miniprogram-render/src/util/tool.js
@@ -1,6 +1,32 @@
 const cache = require('./cache')
 
 /**
+ * 判断是否是内置组件
+ */
+const WX_COMPONENT_MAP = {}
+const WX_COMPONENT_LIST = [
+    'cover-image', 'cover-view', 'match-media', 'movable-area', 'movable-view', 'scroll-view', 'swiper', 'swiper-item', 'view',
+    'icon', 'progress', 'rich-text', 'text',
+    'button', 'checkbox', 'checkbox-group', 'editor', 'form', 'input', 'label', 'picker', 'picker-view', 'picker-view-column', 'radio', 'radio-group', 'slider', 'switch', 'textarea',
+    'functional-page-navigator', 'navigator',
+    'audio', 'camera', 'image', 'live-player', 'live-pusher', 'video', 'voip-room',
+    'map',
+    'canvas',
+    'ad', 'ad-custom', 'official-account', 'open-data', 'web-view',
+    // 特殊补充
+    'capture', 'catch', 'animation'
+]
+WX_COMPONENT_LIST.forEach(name => WX_COMPONENT_MAP[name] = name)
+function checkIsWxComponent(tagName, notNeedPrefix) {
+    const hasPrefix = tagName.indexOf('wx-') === 0
+    if (notNeedPrefix) {
+        return hasPrefix ? WX_COMPONENT_MAP[tagName.slice(3)] : WX_COMPONENT_MAP[tagName]
+    } else {
+        return hasPrefix ? WX_COMPONENT_MAP[tagName.slice(3)] : false
+    }
+}
+
+/**
  * 驼峰转连字符
  */
 function toDash(str) {
@@ -120,6 +146,7 @@ function escapeForHtmlGeneration(value) {
 }
 
 module.exports = {
+    checkIsWxComponent,
     toDash,
     toCamel,
     getId,